+++ src/tests/mod.rs
@@ -1,4 +1,4 @@
 mod lib;
 use lib::*;
 
-mod elementwise;
+mod run;
